requirements.txt
streamlit==1.39.0
azure-cognitiveservices-speech==1.41.1
python-dotenv==1.0.1
pydub==0.25.1